Added findbugs dependency to build; bug fixes # Please enter the commit message for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit. # On branch refactor/findbugs # Changes to be committed: #	modified:   ../../../pom.xml #	modified:   com/pliesveld/flashnote/domain/FlashCard.java #	modified:   com/pliesveld/flashnote/repository/FlashnoteRepositoryImpl.java #	modified:   com/pliesveld/flashnote/schema/DDLExport.java #	modified:   com/pliesveld/flashnote/schema/DropCreateDatabase.java #	modified:   com/pliesveld/flashnote/security/JwtTokenUtil.java #	modified:   com/pliesveld/flashnote/security/UsernameAuditorAware.java #	modified:   com/pliesveld/flashnote/service/StudentService.java #	modified:   com/pliesveld/flashnote/service/StudentServiceImpl.java #	modified:   com/pliesveld/flashnote/web/controller/QuestionBankController.java #	modified:   com/pliesveld/populator/repository/PopulateFlashnoteContainerApplication.java # # Changes not staged for commit: #	modified:   ../../../config/whitespace_only_stylechecks.xml # # Untracked files: #	../../../.gitignore #	../../../.swp #	../../../default_repository/ #	../../../puppy.jpg #	../resources/sql/db-drop.sql #	../resources/sql/db-init.sql #	../../test/resources/scripts/tests/test-data/text/ #

+++ src/main/java/com/pliesveld/flashnote/spring/servlet/SpringWebInitializer.java
@@ -24,7 +24,7 @@ import javax.servlet.ServletRegistration;
 public class SpringWebInitializer extends AbstractAnnotationConfigDispatcherServletInitializer implements WebApplicationInitializer {
     private static final Logger LOG = LogManager.getLogger();
 
-    private int maxUploadSizeInMb = 5 * 1024 * 1024; // 5 MB
+    private long maxUploadSizeInMb = 5 * 1024 * 1024; // 5 MB
 
 
     @Override
@@ -70,7 +70,7 @@ public class SpringWebInitializer extends AbstractAnnotationConfigDispatcherServ
         String uploadDirectory = ""; //ServiceConfiguration.CRM_STORAGE_UPLOADS_DIRECTORY;
         LOG.debug("Configuring MultiPartFilter witn an upload capacity of {}", maxUploadSizeInMb);
 
-        MultipartConfigElement multipartConf = new MultipartConfigElement(uploadDirectory, maxUploadSizeInMb, maxUploadSizeInMb * 2, maxUploadSizeInMb / 2);
+        MultipartConfigElement multipartConf = new MultipartConfigElement(uploadDirectory, maxUploadSizeInMb, maxUploadSizeInMb * 2L, (int) (maxUploadSizeInMb / 2));
         registration.setMultipartConfig(multipartConf);
     }
